pyproject.toml
[tool.poetry]
name = "jetbrains-simple-llm-agent"
version = "0.1.0"
description = "Simple LLM agent for JetBrains internship task"
authors = ["ivanarena <ivan.x28x@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.3"
python-dotenv = "^1.0.1"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
mypy = "^1.12.0"
pylint = "^3.3.1"
langgraph = "^0.2.38"
langchain-google-genai = "^2.0.1"
pydantic = "^2.9.2"
black = "^24.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
